pyproject.toml
[tool.poetry]
name = "pyproger"
version = "0.1.0"
description = "Blog about python "
authors = ["Sergey Vanyushkin <pi3c@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
gunicorn = "^21.2.0"
sqlalchemy = "^2.0.20"
flask-sqlalchemy = "^3.1.1"
flask-security-too = "^5.3.0"
flask-migrate = "^4.0.5"
flask-admin = "^1.6.1"
psycopg2 = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
